+++ CHANGELOG.md
@@ -13,6 +13,9 @@ to 18, which means you'll have to do a dump/restore or use pg_upgrade.
 
 - Allow grouping charts by week and month, instead of just hour and day.
 
+- Add a new JSON export/import, which is more convenient than the CSV export for
+  most use cases.
+
 - Add *Server Management → GeoIP* page to debug GeoIP lookups.
 
 - Password for a new user can now be read from stdin when using `goatcounter db